// src/modules/configuration/model/Form.ts
export interface DataForm {
    IDFormClass: number;
    IDForm: number;
    Name: string;
    Description: string;
    IDEmployee: string;
    State: number;
    UptoDate: Date;
    FormEngine: string;
    FormURLComponent: string;
    HtmlStored: boolean;
    HtmlBodyText: string;
    HtmlExtraText: string;
}